requirements.txt
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
decorator==5.1.1
Django==4.2.1
executing==1.2.0
ExifRead==3.0.0
google-api-core==2.11.0
google-auth==2.19.1
google-cloud-vision==3.4.2
googleapis-common-protos==1.59.0
gpsphoto==2.2.3
grpcio==1.54.2
grpcio-status==1.54.2
idna==3.4
ipython==8.13.2
jedi==0.18.2
matplotlib-inline==0.1.6
mysqlclient==2.1.1
parso==0.8.3
pickleshare==0.7.5
piexif==1.1.3
Pillow==9.5.0
prompt-toolkit==3.0.38
proto-plus==1.22.2
protobuf==4.23.2
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.15.1
python-decouple==3.8
requests==2.31.0
rsa==4.9
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
traitlets==5.9.0
tzdata==2023.3
urllib3==1.26.16
wcwidth==0.2.6
django-sslserver==0.22
